requirements.txt
Django>=3.0,<4.0
psycopg2-binary>=2.8
dj-rest-auth
django-allauth
djangorestframework
djangorestframework-simplejwt
django-cors-headers
gunicorn
django-filter
djangorestframework-filters=1.0.0.dev